Agno_and_Phidata_Apps/requirements.txt
phidata
phidata[all]
agno
python-dotenv
duckduckgo-search
sqlalchemy
pymysql
groq
yfinance
ddgs
wikipedia-api